requirements.txt
faster-whisper>=1.0.0
imageio-ffmpeg>=0.6.0
yt-dlp>=2024.8.6
requests>=2.31.0
PyYAML>=6.0.1
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.0.0
opencc-python-reimplemented>=0.1.7
